added compoent for the wallet component
