+++ src/routes/dashboard/components/WalletKitProvider.svelte
@@ -1,0 +1,43 @@
+<!-- src/lib/components/WalletComponent.svelte -->
+<script>
+  import { onMount } from 'svelte';
+  import { goto } from '$app/navigation';
+  import { walletStore } from '$lib/stores/walletStore';
+  import { fundWithFriendbot } from '$lib/stellar/horizonQueries';
+  import { StellarWalletsKit, WalletNetwork, allowAllModules, XBULL_ID } from '@creit.tech/stellar-wallets-kit';
+
+
+
+  export let buttonText = 'Connect Wallet';
+  const kit = new StellarWalletsKit({
+      network: WalletNetwork.TESTNET,
+      selectedWalletId: XBULL_ID,
+      modules: allowAllModules(),
+  });
+
+  const connectWallet = async () => {
+      try {
+          await kit.openModal({
+              onWalletSelected: async (option) => {
+                  kit.setWallet(option.id);
+                  const { address } = await kit.getAddress();
+                  
+                  if (address) {
+                      await walletStore.registerWithWallet({ publicKey: address });
+                      await fundWithFriendbot(address);
+                      goto('/dashboard');
+                  }
+              }
+          });
+      } catch (error) {
+          console.error('Error connecting wallet:', error);
+      }
+  };
+
+  export { connectWallet };
+</script>
+
+<!-- You can provide a button or any UI elements if needed -->
+<button type="button" class="btn-secondary btn" on:click={connectWallet}>
+  {buttonText}
+</button>
